+++ test/pacts/worldwide_api_pact_test.rb
@@ -6,48 +6,6 @@ describe "GdsApi::Worldwide pact tests" do
 
   let(:api_client) { GdsApi::Worldwide.new(whitehall_api_host) }
 
-  describe "#world_locations" do
-    it "responds with 200 and all world locations" do
-      whitehall_api
-        .given("a world location exists")
-        .upon_receiving("a request to return all world locations")
-        .with(
-          method: :get,
-          path: "/api/world-locations",
-          headers: GdsApi::JsonClient.default_request_headers,
-        )
-        .will_respond_with(
-          status: 200,
-          body: {
-            results: [
-              Pact.like(
-                id: "https://www.gov.uk/api/world-locations/france",
-                title: "France",
-                format: "World location",
-                updated_at: "2020-09-02T06:47:34.000+01:00",
-                web_url: "https://www.gov.uk/world/france",
-                analytics_identifier: "WL1",
-                details: {
-                  slug: "france",
-                  iso2: nil,
-                },
-                organisations: {
-                  id: "https://www.gov.uk/api/world-locations/france/organisations",
-                  web_url: "https://www.gov.uk/world/france#organisations",
-                },
-                content_id: "5e9ecbce-7706-11e4-a3cb-005056011aef",
-              ),
-            ],
-          },
-          headers: {
-            "Content-Type" => "application/json; charset=utf-8",
-          },
-        )
-
-      api_client.world_locations
-    end
-  end
-
   describe "#world_location" do
     describe "#organisations_for_world_location" do
       it "responds with 200 and all worldwide organisations" do