+++ test/worldwide_api_test.rb
@@ -17,26 +17,7 @@ describe GdsApi::Worldwide do
       stub_worldwide_api_has_locations(country_slugs)
 
       response = @api.world_locations
-      assert_equal(country_slugs, response.map { |r| r["details"]["slug"] })
-      assert_equal "Rohan", response["results"][2]["title"]
-    end
-
-    it "should handle the pagination" do
-      country_slugs = (1..50).map { |n| "country-#{n}" }
-      stub_worldwide_api_has_locations(country_slugs)
-
-      response = @api.world_locations
-      assert_equal(
-        country_slugs,
-        response.with_subsequent_pages.map { |r| r["details"]["slug"] },
-      )
-    end
-
-    it "should raise error if endpoint 404s" do
-      stub_request(:get, "#{@base_api_url}/api/world-locations").to_return(status: 404)
-      assert_raises GdsApi::HTTPNotFound do
-        @api.world_locations
-      end
+      assert_equal(country_slugs, response.map { |r| r.dig("details", "slug") })
     end
   end